+++ Shared/NagisaVersion.h
@@ -11,8 +11,8 @@ License: The MIT License
 #define NAGISA_VER
 
 #define NAGISA_VER_MAJOR 0
-#define NAGISA_VER_MINOR 1
-#define NAGISA_VER_BUILD 10
+#define NAGISA_VER_MINOR 2
+#define NAGISA_VER_BUILD 11
 #define NAGISA_VER_REV 0
 #endif